Move stencil shape inference to its own file.
